CAMEL-10804 Create a Salesforce example
This adds configuration for `Camel Salesman` Connected App on Salesforce
developer instance running under `Camelus Bactrianus`
(cbactrianus@gmail.com) user.

Given a Tweet's user uses its `name` as `Last Name` and `screenName` as
custom property `Twitter Screen Name` to create a contact in the
Developer Edition Salesforce instance. The `Twitter Screen Name` field
is defined as unique so in order not to create duplicate Contacts an
update-or-insert operation is performed.