CAMEL-17821 - Provide camel-debug in a profile
Currently, Camel textual debugging requires having camel-debug in
dependencies. This is an extra dependency that we do not want at
runtime.

Previously:

* main java example is configured having it at runtime.
* main xml and yaml doesn't have it

I think a possible good practice is to add the dependency in a profile,
for instance named `camel.debug` which can be activated with the
property camel.debug set to true.

This would be nice to be done for all examples but maybe best to start
with the 3 more simple/important main (java), main-xml and main-yaml to
help starting discussion and see the feedback on it.

I think this is even more important for 3.16 as JMX is exposed by
default when the dependency is available.


Signed-off-by: Aurélien Pupier <apupier@redhat.com>